implemeneted fragment view.
I realized viewmodels should not be shared between 2 activities, so I will implement a separate view with fragments
