+++ app/src/main/java/com/example/top10movies/ui/MoviesActivity.kt
@@ -1,80 +1,21 @@
 package com.example.top10movies.ui
 
-import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.top10movies.adapter.MovieAdapter
-import com.example.top10movies.databinding.ActivityMoviesBinding
+import com.example.top10movies.R
 import com.example.top10movies.repo.MoviesRepo
 import com.example.top10movies.ui.viewmodels.MoviesViewModel
-import com.example.top10movies.util.Constants.Companion.API_KEY
-import com.example.top10movies.util.Constants.Companion.MOVIES_TO_MOVIEDETAILS
-import com.example.top10movies.util.Resource
 
 class MoviesActivity : AppCompatActivity() {
     lateinit var viewModel: MoviesViewModel
-    lateinit var movieAdapter: MovieAdapter
-
-    private lateinit var binding: ActivityMoviesBinding
-
-    val TAG = "Break"
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityMoviesBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        setupRecyclerView()
+        setContentView(R.layout.activity_movies)
 
         val moviesRepo = MoviesRepo()
         val viewModelProviderFactory = MovieViewModelProviderFactory(application, moviesRepo)
         viewModel = ViewModelProvider(this, viewModelProviderFactory)[MoviesViewModel::class.java]
-
-        movieAdapter.setOnItemClickListener { movie ->
-            Toast.makeText(this, "you clicked " + movie.title + "with id: " + movie.id, Toast.LENGTH_SHORT).show()
-        }
-
-        viewModel.movies.observe(this, Observer { response ->
-            when (response) {
-                is Resource.Success -> {
-                    hideLoadingBar()
-                    response.data?.let { movieResponse ->
-                        movieAdapter.differ.submitList(movieResponse.movies)
-                    }
-                }
-                is Resource.Error -> {
-                    hideLoadingBar()
-                    response.message?.let { message ->
-                        Log.e(TAG, "Error: $message")
-                    }
-                }
-                is Resource.Loading -> {
-                    showLoadingBar()
-                }
-            }
-
-        })
     }
-
-    private fun setupRecyclerView() {
-        movieAdapter = MovieAdapter()
-        binding.rvMovieImages.apply {
-            adapter = movieAdapter
-            layoutManager = LinearLayoutManager(this@MoviesActivity)
-        }
-    }
-
-    private fun hideLoadingBar() {
-        // hide loading bar
-        // loadingBar.visibility = View.INVISIBLE
-    }
-
-    private fun showLoadingBar() {
-        // loadingBar.visibility = View.VISIBLE
-    }
-
 }
