+++ app/src/main/java/com/example/top10movies/ui/fragments/MovieDetailsFragment.kt
@@ -1,0 +1,9 @@
+package com.example.top10movies.ui.fragments
+
+import androidx.fragment.app.Fragment
+import com.example.top10movies.R
+
+
+class MovieDetailsFragment : Fragment(R.layout.fragment_movie_details) {
+
+}
