+++ app/src/main/res/layout/activity_movies.xml
@@ -6,12 +6,19 @@
     android:layout_height="match_parent"
     tools:context=".ui.MoviesActivity">
 
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvMovieImages"
+    <FrameLayout
+        android:id="@+id/hostFragment"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:layout_height="match_parent">
+
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/moviesNavHostFragment"
+            android:name="androidx.navigation.fragment.NavHostFragment"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:defaultNavHost="true"
+            app:navGraph="@navigation/movie_nav_graph"/>
+
+    </FrameLayout>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
